+++ .gitignore
@@ -3,7 +3,7 @@
 /cover.txt
 /internal/.git-toplevel
 /internal/.version
-/internal/host/agent_server/agent_server*_*
-/internal/host/agent_server/agent_server*_*.gz
-/internal/host/agent_server_*_*_gz.go
+/internal/host/agent_server_http/agent_server_http_linux_*
+/internal/host/agent_server_http/agent_server_http_linux_*.gz
+/internal/host/agent_server_http_linux_*_gz.go
 /resonance.*.*
