+++ cmd/lib.go
@@ -55,7 +55,7 @@ func wrapHost(ctx context.Context, hst host.Host) (host.Host, error) {
 
 	}
 	if hst.Type() != "localhost" && !optionsMap["disable-agent"] {
-		hst, err = ihost.NewAgent(ctx, hst)
+		hst, err = ihost.NewHttpAgent(ctx, hst)
 		if err != nil {
 			return nil, err
 		}
