+++ internal/host/agent_common_client.go
@@ -1,0 +1,204 @@
+package host
+
+import (
+	"context"
+	"crypto/tls"
+	"fmt"
+	"io"
+	"log/slog"
+	"net"
+	"net/http"
+	"os"
+	"regexp"
+	"strings"
+
+	"go.uber.org/multierr"
+
+	"github.com/fornellas/resonance/host"
+	aNet "github.com/fornellas/resonance/internal/host/agent_server_http/net"
+
+	"github.com/alessio/shellescape"
+	"golang.org/x/net/http2"
+
+	"github.com/fornellas/resonance/log"
+)
+
+func getGoArch(machine string) (string, error) {
+	matched, err := regexp.MatchString("^i[23456]86$", machine)
+	if err != nil {
+		panic(err)
+	}
+	if matched {
+		return "386", nil
+	}
+	matched, err = regexp.MatchString("^x86_64$", machine)
+	if err != nil {
+		panic(err)
+	}
+	if matched {
+		return "amd64", nil
+	}
+	matched, err = regexp.MatchString("^armv6l|armv7l$", machine)
+	if err != nil {
+		panic(err)
+	}
+	if matched {
+		return "arm", nil
+	}
+	matched, err = regexp.MatchString("^aarch64$", machine)
+	if err != nil {
+		panic(err)
+	}
+	if matched {
+		return "arm64", nil
+	}
+	return "", fmt.Errorf("machine %#v not supported by agent", machine)
+}
+
+func getAgentBinGz(ctx context.Context, hst host.Host) ([]byte, error) {
+	cmd := host.Cmd{
+		Path: "uname",
+		Args: []string{"-m"},
+	}
+	waitStatus, stdout, stderr, err := host.Run(ctx, hst, cmd)
+	if err != nil {
+		return nil, err
+	}
+	if !waitStatus.Success() {
+		return nil, fmt.Errorf(
+			"failed to run %s: %s\nstdout:\n%s\nstderr:\n%s",
+			cmd, waitStatus.String(), stdout, stderr,
+		)
+	}
+	goarch, err := getGoArch(strings.TrimRight(stdout, "\n"))
+	if err != nil {
+		return nil, err
+	}
+	osArch := fmt.Sprintf("linux.%s", goarch)
+
+	agentBinGz, ok := AgentHttpBinGz[osArch]
+	if !ok {
+		return nil, fmt.Errorf("%s not supported by agent", osArch)
+	}
+	return agentBinGz, nil
+}
+
+func getTmpFile(ctx context.Context, hst host.Host, template string) (string, error) {
+	cmd := host.Cmd{
+		Path: "mktemp",
+		Args: []string{"-t", fmt.Sprintf("%s.XXXXXXXX", template)},
+	}
+	waitStatus, stdout, stderr, err := host.Run(ctx, hst, cmd)
+	if err != nil {
+		return "", err
+	}
+	if !waitStatus.Success() {
+		return "", fmt.Errorf(
+			"failed to run %s: %s\nstdout:\n%s\nstderr:\n%s",
+			cmd, waitStatus.String(), stdout, stderr,
+		)
+	}
+	return strings.TrimRight(stdout, "\n"), nil
+}
+
+func copyReader(ctx context.Context, hst host.Host, reader io.Reader, path string) error {
+	cmd := host.Cmd{
+		Path:  "sh",
+		Args:  []string{"-c", fmt.Sprintf("cat > %s", shellescape.Quote(path))},
+		Stdin: reader,
+	}
+	waitStatus, stdout, stderr, err := host.Run(ctx, hst, cmd)
+	if err != nil {
+		return err
+	}
+	if !waitStatus.Success() {
+		return fmt.Errorf(
+			"failed to run %s: %s\nstdout:\n%s\nstderr:\n%s",
+			cmd, waitStatus.String(), stdout, stderr,
+		)
+	}
+	return nil
+}
+
+type writerLogger struct {
+	Logger *slog.Logger
+}
+
+func (wl writerLogger) Write(b []byte) (int, error) {
+	lines := strings.Split(string(b), "\n")
+	for i, line := range lines {
+		if len(line) == 0 && i+1 == len(lines) {
+			break
+		}
+		wl.Logger.Error("Agent", "line", line)
+	}
+	return len(b), nil
+}
+
+func (a *AgentHttpClient) spawn(ctx context.Context) error {
+	logger := log.MustLogger(ctx)
+
+	stdinReader, stdinWriter, err := os.Pipe()
+	if err != nil {
+		return err
+	}
+
+	stdoutReader, stdoutWriter, err := os.Pipe()
+	if err != nil {
+		return err
+	}
+
+	a.Client = &http.Client{
+		Transport: &http2.Transport{
+			DialTLSContext: func(
+				ctx context.Context, network, addr string, cfg *tls.Config,
+			) (net.Conn, error) {
+				return aNet.Conn{
+					Reader: stdoutReader,
+					Writer: stdinWriter,
+				}, nil
+			},
+			AllowHTTP: true,
+		},
+	}
+
+	go func() {
+		defer func() { a.waitCn <- struct{}{} }()
+		waitStatus, err := a.Host.Run(ctx, host.Cmd{
+			Path:   a.path,
+			Stdin:  stdinReader,
+			Stdout: stdoutWriter,
+			Stderr: writerLogger{
+				Logger: logger,
+			},
+		})
+		if err != nil {
+			logger.Error("failed to run agent", "err", err)
+		}
+		if !waitStatus.Success() {
+			logger.Error("agent exited with error", "error", waitStatus)
+		}
+		stdinWriter.Close()
+		stdoutReader.Close()
+		stdinReader.Close()
+		stdoutWriter.Close()
+	}()
+
+	resp, err := a.get("/ping")
+	if err != nil {
+		return multierr.Combine(err, a.Close())
+	}
+
+	bodyBytes, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return multierr.Combine(err, a.Close())
+	}
+	if string(bodyBytes) != "Pong" {
+		return multierr.Combine(
+			fmt.Errorf("pinging agent failed: unexpected body %#v", string(bodyBytes)),
+			a.Close(),
+		)
+	}
+
+	return nil
+}
